fix: build error - root dir is expected to contain all source files
