+++ generator/lib/ArtifactProviders/ExecutableSeed/index.ts
@@ -1,7 +1,6 @@
 import * as Babel from '@babel/core'
-import { Index } from '~/src/utils'
-import { TemplateInfo } from '../../templates'
-import { datasourceUrlEnvironmentVariableName } from '../../utils'
+import { TemplateInfo } from '~/src/templates'
+import { datasourceUrlEnvironmentVariableName, Index } from '~/src/utils'
 import { ArtifactProvider } from '../types'
 import { babelPluginTransformTemplate } from './babel-plugin-transform-template'
 import { File } from '~/src/types'
